+++ src/components/DateInput.tsx
@@ -1,4 +1,4 @@
-import { createSignal, For, Match, Show, Switch } from "solid-js"
+import { createEffect, createMemo, createSignal, For, Match, Show, Switch } from "solid-js"
 import { FormComponentBase } from "../FormType"
 
 const DateInput: FormComponentBase = props => {
@@ -14,6 +14,27 @@ const DateInput: FormComponentBase = props => {
 
   const [disableClickRemark] = createSignal((config.formMode > 2  && props.comments == 0 ) ? true : false);
   const [enableRemark] = createSignal(props.component.enableRemark !== undefined ? props.component.enableRemark : true );
+
+  let today = new Date();
+  let dd = String(today.getDate());
+  let mm = String(today.getMonth() + 1); //January is 0!
+  let yyyy = String(today.getFullYear());
+  if(Number(dd) < 10){
+      dd = '0' + dd
+  } 
+  if(Number(mm) < 10){
+      mm = '0' + mm
+  } 
+  
+  let todayDate = yyyy + '-' + mm + '-' + dd;
+
+  let minDate : any, maxDate: any;
+  createMemo(() => {
+    if(props.component.rangeInput){
+      minDate = (props.component.rangeInput[0].min !== undefined ) ? (props.component.rangeInput[0].min === 'today') ? todayDate : props.component.rangeInput[0].min : '';
+      maxDate = (props.component.rangeInput[0].max !== undefined ) ? (props.component.rangeInput[0].max === 'today') ? todayDate : props.component.rangeInput[0].max : '';
+    }
+  })
 
   return (
 		<div class="md:grid md:grid-cols-3 border-b border-gray-300/[.40] dark:border-gray-200/[.10] p-2">
@@ -56,7 +77,9 @@ const DateInput: FormComponentBase = props => {
                 disabled = { disableInput() }
                 onChange={(e) => {
                   props.onValueChange(e.currentTarget.value);
-                } }
+                } }            
+                min = {minDate}    
+                max = {maxDate}
           />
           <Show when={props.validationMessage.length > 0}>
             <For each={props.validationMessage}>